wip: switch to Turso database
* Work in progress switch to Turso database.

+++ testing/src/main/java/com/puckowski/testing/dto/TestCaseDTO.java
@@ -2,4 +2,4 @@ package com.puckowski.testing.dto;
 
 import java.time.LocalDateTime;
 
-public record TestCaseDTO(Long id, Long testPlanId, String name, String description, String status, LocalDateTime createdAt, String expectedResult, String priority, String steps) {}
+public record TestCaseDTO(Integer id, Integer testPlanId, String name, String description, String status, LocalDateTime createdAt, String expectedResult, String priority, String steps) {}
