+++ testing/src/main/java/com/puckowski/testing/dto/TestPlanCountDTO.java
@@ -3,4 +3,4 @@ package com.puckowski.testing.dto;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public record TestPlanCountDTO(Long count) {}
+public record TestPlanCountDTO(Integer count) {}
