+++ testing/src/main/java/com/puckowski/testing/dto/TestPlanDTO.java
@@ -3,5 +3,7 @@ package com.puckowski.testing.dto;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public record TestPlanDTO(Long id, String name, String description, LocalDateTime createdAt, String status, List<TestTagDTO> tagList,
+public record TestPlanDTO(Integer id, String name,
+                          String description, LocalDateTime createdAt,
+                          String status, List<TestTagDTO> tagList,
                           List<TestCaseDTO> testCases) {}
